projects/04-llamaguard/requirements.txt
# Core ML/DL frameworks
torch>=2.7.0
transformers>=4.39.0
accelerate>=0.20.0

# Vector database and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
numpy>=1.24.0

# LangChain and LangGraph
langchain-core>=0.1.0
langchain-upstage>=0.1.0
langgraph>=0.0.1
pydantic>=2.0.0

# External API clients
openai>=1.52.2
tavily-python>=0.3.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
